Termindo el proceso de evaluar falta hacer la nota media del alumno y mostrarla y hacer el visualizar ganadores, tambien falta comentar las apis y hacer un text

+++ Backend/src/main/java/com/andaluciaskills/andaluciasckills/Mapper/EvaluacionItemMapper.java
@@ -17,7 +17,7 @@ public class EvaluacionItemMapper implements GenericMapper<EvaluacionItem, DtoEv
         
         DtoEvaluacionItem dto = new DtoEvaluacionItem();
         dto.setIdEvaluacionItem(entity.getIdEvaluacionItem());
-        dto.setValoracion(entity.getValoracion());
+        dto.setValoracion(entity.getValoracion().doubleValue());
         dto.setEvaluacion_idEvaluacion(entity.getEvaluacion_idEvaluacion());
         dto.setItem_idItem(entity.getItem_idItem());
         return dto;
